pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "allie-flowkit"
version = "0.1.dev0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<4"
license = { file = "LICENSE" }
authors = [{ name = "ANSYS, Inc.", email = "pyansys.core@ansys.com" }]
maintainers = [{ name = "ANSYS, Inc.", email = "pyansys.core@ansys.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi >= 0.111.1,<1",
    "langchain >= 0.2.11,<1",
    "pydantic >= 2.8.2,<3",
    "pymupdf >= 1.24.9,<2",
    "python_pptx >= 0.6.23,<1",
    "PyYAML >= 6.0.1,<7",
]

[project.optional-dependencies]
    all = [
        "uvicorn[standard] >= 0.30.5,<1",
    ]
[project.scripts]
allie-flowkit = "allie.flowkit.__main__:main"


[tool.flit.module]
name = "allie.flowkit"

[tool.codespell]
skip = "*.pyc,*.xml,*.txt,*.gif,*.png,*.jpg,*.js,*.html,*.doctree,*.ttf,*.woff,*.woff2,*.eot,*.mp4,*.inv,*.pickle,*.ipynb,flycheck*,./.git/*,./.hypothesis/*,*.yml,./docs/build/*,./docs/images/*,./dist/*,*~,.hypothesis*,./docs/source/examples/*,*cover,*.dat,*.mac"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 100
default_section = "THIRDPARTY"
src_paths = ["doc", "src", "tests"]

[tool.pydocstyle]
convention = "numpy"

[tool.mypy]
strict = true
ignore_missing_imports = true
